backend/requirements.txt
# ============================================================
# Consultancy Intelligence Agent – Python Dependencies
# ============================================================
# All packages are free and open-source.
# Compatible with Python 3.10+

# --- Web Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0

# --- AI / ML ---
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
langchain==0.3.14
langchain-community==0.3.14
langchain-huggingface==0.1.2
transformers==4.47.1
torch==2.5.1

# --- Data & Configuration ---
pydantic==2.10.4
pydantic-settings==2.7.1
pandas==2.2.3
python-dotenv==1.0.1

# --- HTTP Client (Slack integration) ---
httpx==0.28.1
